Class5/HW2.1/RAG_W_SQLite_FAISS/requirements.txt
requests
arxiv
PyMuPDF
pytesseract
Pillow
sentence-transformers
faiss-cpu
numpy
openai
tqdm
